Code/requirements.txt
numpy>=1.23.0
scikit-learn>=1.1.0
pillow>=9.0.0
torch>=2.0.0
torchvision>=0.15.0
